Move sentinel DNS computation into `StubResolver`
